requirements.txt
Django==4.2.20
qrcode==7.4.2
Pillow==11.0.0
python-dateutil
